requirements.txt
appdirs==1.4.4
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Faker==5.0.2
idna==2.10
Jinja2==2.11.2
Markdown==3.3.3
MarkupSafe==1.1.1
mypy-extensions==0.4.3
nested-lookup==0.2.21
pathspec==0.8.1
python-dateutil==2.8.1
python-slugify==4.0.1
PyYAML==5.3.1
regex==2020.11.13
requests==2.25.1
requests-cache==0.5.2
six==1.15.0
text-unidecode==1.3
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
xmltodict==0.12.0
